Refactor favorpage tests to prevent conflicts during testing
Unify test utils
