+++ app/src/sharedTestUtils/java/ch/epfl/favo/TestUtils.java
@@ -11,6 +11,7 @@ import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
 
+import java.util.Random;
 import java.util.concurrent.ExecutionException;
 
 import ch.epfl.favo.util.DependencyFactory;
@@ -80,5 +81,16 @@ public class TestUtils {
             && view.equals(((ViewGroup) parent).getChildAt(position));
       }
     };
+
+  }
+  public static String generateRandomString(int targetStringLength) {
+    // String title = "sample_favor";
+    int leftLimit = 97; // letter 'a'
+    int rightLimit = 122; // letter 'z'
+    return new Random()
+            .ints(leftLimit, rightLimit + 1)
+            .limit(targetStringLength)
+            .collect(StringBuilder::new, StringBuilder::appendCodePoint, StringBuilder::append)
+            .toString();
   }
 }
